// ArchiCore/src/app/api-client/model/missionCreateRequest.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface MissionCreateRequest { 
    codeMission: string;
    exercice: string;
    codeDossier: string;
}