// ArchiCore/src/app/api-client/model/missionUpdateRequest.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface MissionUpdateRequest { 
    statut?: MissionUpdateRequest.StatutEnum;
    exercice?: string;
}
export namespace MissionUpdateRequest {
    export const StatutEnum = {
        Planifiee: 'PLANIFIEE',
        EnCours: 'EN_COURS',
        Terminee: 'TERMINEE',
        Suspendue: 'SUSPENDUE'
    } as const;
    export type StatutEnum = typeof StatutEnum[keyof typeof StatutEnum];
}